Prevent potential serialization bug, display message if recipient of /hcpay is online.

+++ src/main/java/regalowl/hyperconomy/util/MessageBuilder.java
@@ -7,7 +7,7 @@ public class MessageBuilder {
 	private transient LanguageFile L;
 	private String message;
 	
-	private int amount;
+	private double amount;
 	private double price;
 	private String objectName;
 	private String playerName;
@@ -32,7 +32,7 @@ public class MessageBuilder {
 		this.message = message;
 	}
 
-	public void setAmount(int amount) {
+	public void setAmount(double amount) {
 		this.amount = amount;
 	}
